requirements.txt
requests
flask
python-dotenv
requests-oauthlib
